Comments and Summary
I added comments to the code that help me understand how it works in more plain english. In a prod env. I might just store these myself and not commit them. I also put a ton of comments in the test file that should drive how I test this to make sure I understand how it works. The current task will be to write tests that pass for how the code is currently written but that fail if the current code doesn't line up with the shared logic in the readme.