Refactored to Switch
So first step is I gutted the function and rewrote it as a simple switch statement. This is not the best way to solve this but its something I'd expect to see in some solutions. The thing I would focus on is that this actually works. Its pretty clean, functional, its very easy to extend and switch statements are pretty performant. They optimize better than ifs and each items logic is self contained.